requirements.txt
easyocr
opencv-python>=4.13.0.92
numpy>=1.24,<2.0
python-dateutil
PyYAML
